+++ src/library/tests/text_export.rs
@@ -1,0 +1,26 @@
+use std::io::BufReader;
+
+use grebedb::{Database, Options};
+
+#[test]
+fn test_export() {
+    let mut database = Database::open_memory(Options::default()).unwrap();
+
+    database.put("key1", "value1").unwrap();
+    database.put("key2", "value2").unwrap();
+    database.put("key3", "value3").unwrap();
+
+    let mut file = Vec::new();
+
+    grebedb::export::export(database, &mut file, |_| {}).unwrap();
+
+    let database = Database::open_memory(Options::default()).unwrap();
+
+    let (mut database, _) =
+        grebedb::export::import(database, BufReader::new(std::io::Cursor::new(file)), |_| {})
+            .unwrap();
+
+    assert_eq!(database.get("key1").unwrap(), Some(b"value1".to_vec()));
+    assert_eq!(database.get("key2").unwrap(), Some(b"value2".to_vec()));
+    assert_eq!(database.get("key3").unwrap(), Some(b"value3".to_vec()));
+}